Only publish stream events when sendStreamEvents is set.
The publishing of stream events is only enabled when the
"sendStreamEvents' configuration is set as follows:

[plugin "events-kafka"]
   sendStreamEvents=true

Note that this constitutes a breaking change compared to the previous
stable version (stable-3.4), in which stream events were published by
default.

This is because multi-site now already publishes stream events
[see Iafe5a8155] by binding directly the StreamEventsPublisher from the
events-broker library. Thus direct publishing from events-kafka would be
redundant and cause twice as many messages to be published.

Bug: Issue 14910
Change-Id: Ie77ad7d10fe963568499c17aa6faa512006e30db