Abstract Publisher/Subscriber into generic interfaces
Decouple the interface to send/receive messages to Kafka
using generic interfaces bound to the actual native implementation
classes.

The KafkaSession does not refer anymore to a native KafkaProducer
allowing other implementations to be plugged in.

This is a preparation work to introduce a REST-API client
based access to Kafka REST Proxy through parameters on
the events broker plugin config.

Change-Id: I6034915c5538e3df365a45e2f134bab50aff932f