Abstract Session/Subscriber into generic interfaces
Decouple the interface to send/receive messages to Kafka
using generic interfaces bound to the actual native implementation
classes.

This is a preparation work to introduce a REST-API client
based access to Kafka REST Proxy through parameters on
the events broker plugin config.

The native implementation using Kafka clients is the only
binding currently available for KafkaSession and KafkaEventSubscriber.

Change-Id: I6034915c5538e3df365a45e2f134bab50aff932f