Kafka REST Client: avoid clushes between clients
When multiple receivers are listening to different topics
they need to avoid clashing by using different group ids.

With REST Client API based receivers, two receivers on the
same group would deadlock with each other.

Also set explicitly the auto-commit and receive timeout
so that receivers are not considered as failed by the
the REST Proxy and does not start the 5 mins rebalancing
loop.

Change-Id: I774c295276441365b8939e36ffd93fbcd8fd60bc