Kafka REST Client: avoid clashes between clients
When multiple receivers are listening to different topics
they need to avoid clashing by using different group ids.

With REST Client API based receivers, two receivers on the
same group would deadlock with each other.

Also set explicitly the auto-commit and consumer.request.timeout.ms,
the receive loops are blocking in the Kafka REST Proxy.

Staying in a receive loop too much is risky, because if the Gerrit
receiver becomes unresponsive or disconnects for whatever reason,
the receiver would be considered as failed and would start a
5 mins rebalancing loop.

Change-Id: I774c295276441365b8939e36ffd93fbcd8fd60bc