Use Kafka REST Proxy id to subscribe to the correct instance
When using a Kafka REST Proxy with multiple instances behind a
workload balanceer, the consumers are allocated on one of the
allocated instances. The Kafka REST client needs to understand
which one is the instance that the client should talk to.

For deployments that can use workload balancers with sticky
sessions, there isn't any logic needed on the client side, because
the cookies allocated and stored on the client would do the job.
However, when using a standard K8s deployment, the sticky session
allocation isn't an option and any of the incoming traffic may
reach any of the pods.

Kafka REST Proxy allows with this change to specify the REST_PROXY_ID
with the placeholder ${rest_proxy_id} in the API URL, to help
redirecting the calls to the correct instance.

Example:
[plugin "events-kafka"]
  restApiUri = http://kafka-rest:8080/${kafka_rest_id}

Change-Id: I396c3bc759ee05547e8c16cb4806b4fea63b833c