Use Kafka REST Proxy id to subscribe to the correct instance
Kafka REST Proxy allows to add id parameter to help recognise the
instance. Append this parameter to the subscription URL to make sure
that all calls are redirected by load balancer to the correct instance.
This allows to workaround situation when sticky session are not allowed
on the load balancer level

Change-Id: I396c3bc759ee05547e8c16cb4806b4fea63b833c